requirements.txt
requests=2.14.2
instabot==0.4.5
python-dotenv==0.9.1
Pillow==5.3.0